+++ python/paddle/incubate/distributed/fleet/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
+# Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,4 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-__all__ = []
+from paddle.distributed.fleet.recompute import recompute_sequential, recompute_hybrid
+
+__all__ = ["recompute_sequential", "recompute_hybrid"]
